Publish now determines erts vsn based on beam info
Now using beam lib to discover the compiler version a beam was compiled for.
That information is then translated into an Erts vsn.  Currently this resolution is
done stically. There is a stub function in place that will be a fallback in case the
static function does not work.  The idea is to search the repos for the compiler
version in question and then return the highest erts version it is found under. This is
not currently implemented.

*Note* the resolution will not work on stripped beams.  This code introduces a new requirement:
No stripped beams in erlware repos.