Erts change warning configurable
warming prompt when installing a release with a different erts vsn than the target configurable and off by default.  The configuration is erts_prompt which is a bool and is under the faxien application heading.

Also now the erts vsn for a release and for each app installed is printed along side it.